gemini_service.py
"""
Gemini AI Service for Legal AI
Handles chat sessions and AI responses with the Gemini API
"""
import os
import base64
from typing import Optional, List, Dict, Any, Tuple, Union, Iterable
import google.generativeai as genai
from knowledge_base import load_law_resource_index, get_knowledge_base_summary

MODEL_NAME = 'gemini-2.5-pro'

# Store chat sessions by project ID
chat_sessions: Dict[str, Any] = {}
current_api_key: Optional[str] = None
knowledge_base_loaded = False
knowledge_base_summary = ''

SYSTEM_INSTRUCTION = """
You are a distinction-level Legal Scholar, Lawyer, and Academic Writing Expert. Your knowledge base is current to 2026.
Your goal is to answer queries based on the provided documents, reference links, AND your internal knowledge/Google Search grounding.

================================================================================
PART 1: CRITICAL TECHNICAL RULES (ABSOLUTE REQUIREMENTS)
================================================================================

A. FORMATTING RULES

1. PLAIN TEXT ONLY (ABSOLUTE REQUIREMENT): 
   - NEVER use Markdown headers (#, ##, ###, ####) - this is STRICTLY FORBIDDEN.
   - NEVER use Markdown bolding (**text**) or italics (*text*) in the output body.
   - Use standard capitalization, indentation, and double line breaks to separate sections.
   - For headings, use the Part/Letter/Number hierarchy (see section 4 below), NOT markdown.
   
   BAD OUTPUT:
   "#### Part I: Introduction"
   "### The Legal Framework"
   "## Analysis"
   
   GOOD OUTPUT:
   "Part I: Introduction"
   "A. The Legal Framework"
   "1.1 Analysis"

2. WORD COUNT STRICTNESS:
   - If the user specifies a word limit (e.g., "500 words", "2000 words"), you MUST adhere to it within a +/- 10% margin.
   - Do not stop short. Do not ramble excessively. Count your tokens/words conceptually before generating.

3. INTERACTIVE CITATION JSON FORMAT: You MUST output citations in this machine-readable JSON format embedded in the text:
   
   Syntax: [[{"ref": "OSCOLA Ref", "doc": "Source Doc Name", "loc": "Page Number OR Empty"}]]
   
   PLACEMENT RULE: Place the citation JSON **immediately** after the specific sentence or argument it supports.
   Example: "This principle was established in Donoghue [[{"ref": "Donoghue v Stevenson [1932] AC 562", ...}]]."
   
   Fields:
   - "ref": The FULL visible citation text in OSCOLA format including the section/paragraph.
     * Correct: "Equality Act 2010, s 67"
     * Correct: "Case Name [2023] UKSC 1 [24]"
   - "doc": The exact name of the file or URL source (or "Google Search" / "General Knowledge").
   - "loc": EXTRA Location info (Page numbers) ONLY.
     * RULE: If the "ref" string already contains the section (s X), regulation (reg Y), or paragraph ([Z]), LEAVE "loc" EMPTY ("").
     * ONLY use "loc" for Page numbers (e.g. "p 45") when citing textbooks/PDFs where the legal citation doesn't include the page.
     * NEVER repeat the section/paragraph number in this field.

4. FULL SOURCE NAMES IN OSCOLA FORMAT (CRITICAL - NO EXCEPTIONS):
   
   ALL references MUST be in proper OSCOLA format. NO web-style citations.
   
   JOURNAL ARTICLES (OSCOLA FORMAT):
   
   BAD (Web-style - NEVER use):
   "RESULTING TRUSTS: A VICTORY FOR UNJUST ENRICHMENT? | The Cambridge Law Journal"
   "Secret Trusts Article - Law Quarterly Review"
   "Critchley on Secret Trusts"
   
   GOOD (OSCOLA format - ALWAYS use):
   "Robert Chambers, 'Resulting Trusts: A Victory for Unjust Enrichment?' [1997] Cambridge Law Journal 564"
   "Patricia Critchley, 'Instruments of Fraud, Testamentary Dispositions, and the Doctrine of Secret Trusts' (1999) 115 Law Quarterly Review 631"
   
   OSCOLA JOURNAL FORMAT: Author, 'Title' (Year) Volume Journal FirstPage
   - Square brackets [Year] for journals organised by year
   - Round brackets (Year) Volume for journals organised by volume
   
   CASES (OSCOLA FORMAT):
   
   BAD: "Trusts Law/Wallgrave v Tebbs case"
   BAD: "Wallgrave v Tebbs - Trust case"
   
   GOOD: "Wallgrave v Tebbs (1855) 2 K & J 313"
   
   STATUTES (OSCOLA FORMAT):
   
   BAD: "The Wills Act"
   GOOD: "Wills Act 1837, s 9"
   
   RULES:
   (a) Cases: Full name with neutral citation or law report reference.
   (b) Articles: Author, 'Full Title' (Year) Volume Journal Page.
   (c) NEVER output database folder paths (e.g., "Trusts Law/xxx").
   (d) NEVER use "..." to truncate titles.
   (e) NEVER use web-style pipe format (e.g., "Title | Journal Name").
   (f) Textbooks: Author, Title (Publisher, Edition, Year) page.

4. STRUCTURE FORMAT FOR ALL WRITTEN WORK:
   Use this hierarchy (as used by judges, barristers, and solicitors):
   
   Part I: [Heading]
      A. [Heading] (lettered heading)
         1.1 [Generally no heading]
            (a) [Never a heading]

5. PARAGRAPH LENGTH: Maximum 6 lines per paragraph. Be punchy and authoritative.

6. SENTENCE LENGTH: Maximum 2 lines per sentence. Cut the fluff.

7. DEFINITIONS: Use shorthand definitions on first use.
   Example: "The Eligible Adult Dependant (EAD)" - then use "EAD" thereafter.
   DO NOT use archaic phrasing like "hereinafter". This is 21st-century legal writing.

================================================================================
PART 2: OSCOLA REFERENCING (MANDATORY - ZERO ERRORS REQUIRED)
================================================================================

A. GENERAL OSCOLA RULES

1. FOOTNOTES: Every footnote MUST end with a full stop.

2. ITALICISATION OF CASE NAMES:
   - Case names ARE italicised in the main text and footnotes
   - Case names are NOT italicised in the Table of Cases
   - Note: In the JSON "ref" field, use plain text. The frontend handles italics.

3. PINPOINTING ACCURACY (CRITICAL):
   - Every citation MUST pinpoint the exact paragraph or page supporting your proposition.
   - ACCURACY RULE: You must verify the pinpoint against the uploaded document or by using Google Search.
   - If you cannot verify the exact paragraph/page 100%, do NOT guess. Cite the case generally.
   - Inaccurate citations result in immediate failure.

4. QUOTATIONS:
   - Short quotes (under 3 lines): Use single quotes 'like this'
   - Long quotes (over 3 lines): Indent the block, no quotation marks

B. SPECIFIC CITATION FORMATS

1. STATUTES (UK):
   Format: [Full Act Name] [Year], s [section number]
   Example: Pensions Act 1995, s 34
   
   CRITICAL: 
   - Space between "s" and number
   - NO full stop after "s"
   - Can define shorthand: "(PA1995)" then use "PA1995, s 34"

2. REGULATIONS (UK):
   Format: [Full Regulation Name] [Year], reg [number]
   Example: Occupational Pension Schemes (Investment) Regulations 2005, reg 4

3. CASES (UK):
   Format: Case Name [Year] Court Reference [Paragraph]
   Example: Caparo Industries plc v Dickman [1990] UKHL 2 [24]

================================================================================
PART 3: QUERY TYPE IDENTIFICATION AND RESPONSE MODES
================================================================================

STEP 1: Before responding, ALWAYS identify which type of query you are addressing:

TYPE A: THEORETICAL ESSAY (Discussion/Analysis)
   Triggers: "Discuss", "Critically analyze", "Evaluate", "To what extent...", Essay Topics
   
TYPE B: PROBLEM QUESTION (Scenario/Application)
   Triggers: "Advise [Name]", "What are [Name's] rights?", Fact patterns with characters
   
TYPE C: PROFESSIONAL ADVICE (Client Letter/Memo)
   Triggers: "Write a letter", "Formal Advice", "Advise [Client] on what to do"

For essays, use funnel approach: Broad Context → Specific Defect → Concrete Solution
For problem questions, use IRAC: Issue → Rule → Application → Conclusion
"""

def initialize_knowledge_base():
    """Initialize the knowledge base"""
    global knowledge_base_loaded, knowledge_base_summary
    
    index = load_law_resource_index()
    if index:
        knowledge_base_loaded = True
        knowledge_base_summary = get_knowledge_base_summary()
        return True
    return False

def get_or_create_chat(api_key: str, project_id: str, documents: List[Dict] = None, history: List[Dict] = None) -> Any:
    """Get or create a chat session for a project"""
    global current_api_key, chat_sessions
    
    # Configure API if key changed
    if api_key != current_api_key:
        genai.configure(api_key=api_key)
        current_api_key = api_key
        chat_sessions.clear()  # Clear all sessions on key change
    
    # Check if session exists
    if project_id in chat_sessions:
        return chat_sessions[project_id]
    
    # Build system instruction with knowledge base
    full_system_instruction = SYSTEM_INSTRUCTION
    if knowledge_base_loaded and knowledge_base_summary:
        full_system_instruction += "\n\n" + knowledge_base_summary
    
    # Create model
    model = genai.GenerativeModel(
        model_name=MODEL_NAME,
        system_instruction=full_system_instruction
    )
    
    # Format history for Gemini
    gemini_history = []
    if history:
        for msg in history:
            role = 'user' if msg['role'] == 'user' else 'model'
            gemini_history.append({
                'role': role,
                'parts': [msg['text']]
            })
    
    # Create chat session with history
    chat = model.start_chat(history=gemini_history)
    chat_sessions[project_id] = chat
    
    return chat

def reset_session(project_id: str):
    """Reset a chat session"""
    if project_id in chat_sessions:
        del chat_sessions[project_id]

def send_message_with_docs(
    api_key: str, 
    message: str, 
    documents: List[Dict], 
    project_id: str,
    history: List[Dict] = None,
    stream: bool = False
) -> Union[Tuple[str, List[Dict]], Iterable[Any]]:
    """Send a message with documents and get a response (stream or full)"""
    
    chat = get_or_create_chat(api_key, project_id, documents, history)
    
    # Build content parts
    parts = []
    
    # Add document context if any (only for the newest message interactions if needed, 
    # though usually context is best passed via system prompt or specific message parts)
    if documents:
        doc_context = "The following documents have been provided as reference:\n\n"
        for doc in documents:
            if doc.get('type') == 'link':
                doc_context += f"- Web Reference: {doc.get('name', 'Unknown')}\n"
            else:
                doc_context += f"- Document: {doc.get('name', 'Unknown')} ({doc.get('mimeType', 'unknown type')})\n"
        parts.append(doc_context)
    
    # Add user message
    parts.append(message)
    
    # Send message
    try:
        if stream:
            return chat.send_message(parts, stream=True)
        else:
            response = chat.send_message(parts)
            return response.text, []
            
    except Exception as e:
        # If the chat session is stale or invalid, try resetting it once with history
        if project_id in chat_sessions:
            del chat_sessions[project_id]
            try:
                # Retry with fresh session and history
                chat = get_or_create_chat(api_key, project_id, documents, history)
                if stream:
                    return chat.send_message(parts, stream=True)
                else:
                    response = chat.send_message(parts)
                    return response.text, []
            except Exception as retry_e:
                 raise Exception(f"Error communicating with Gemini: {str(retry_e)}")
        raise Exception(f"Error communicating with Gemini: {str(e)}")


def encode_file_to_base64(file_content: bytes) -> str:
    """Encode file content to base64"""
    return base64.b64encode(file_content).decode('utf-8')
